update delta3d.py from delta3d, fixed a ton of warnings, and updated some of the code based on the minor changes to dtAgeiaPhysX.

+++ SimulationCore/trunk/source/SimCore/Components/WeatherComponent.cpp
@@ -441,7 +441,7 @@ namespace SimCore
             UpdateFog();
 
             // Change the Cloud Coverage
-            mEphemerisEnvironmentActor->ChangeClouds((atmosActor->GetCloudCoverage() / 10), 
+            mEphemerisEnvironmentActor->ChangeClouds(int(atmosActor->GetCloudCoverage() / 10), 
                                                       atmosActor->GetWindSpeedX(), 
                                                       atmosActor->GetWindSpeedY());
             // Change Wind
